rebase and lint fixes

fix sts v1 dependency

+++ apis/eks/v1alpha1/nodegroup_types.go
@@ -210,7 +210,7 @@ type NodeGroupScalingConfig struct {
 // NodeGroupScalingConfigStatus is the observed scaling configuration for a node group.
 type NodeGroupScalingConfigStatus struct {
 	// The current number of worker nodes for the managed node group.
-	DesiredSize *int64 `json:"desiredSize,omitempty"`
+	DesiredSize *int32 `json:"desiredSize,omitempty"`
 }
 
 // NodeGroupObservation is the observed state of a NodeGroup.
